Rename module FuzzySortedDict => fuzzysorteddict
Build tools (poetry) are picky about capitalization
